llama-server/requirements.txt
Flask==2.3.2
llama_cpp_python==0.1.78
